refactor: centralize axios import
