+++ frontend/src/components/AuthWidget.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../lib/axios';
 import { toast } from 'react-toastify';
 import { motion } from 'framer-motion';
