+++ frontend/src/components/dashboard/DashboardWidget.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, screen, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../../lib/axios';
 import DashboardWidget from './DashboardWidget';
 import { toast } from 'react-toastify';
