+++ frontend/src/components/designs/DesignWidget.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { Link } from 'react-router-dom';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../../lib/axios';
 import { toast } from 'react-toastify';
 import Slider from "react-slick";
 import { motion } from 'framer-motion';
