+++ frontend/src/components/main/InteractiveGuide.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render, fireEvent, waitFor, screen } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../../lib/axios';
 import InteractiveGuide from './InteractiveGuide';
 
 // ماک کردن ماژول axios
