+++ frontend/src/lib/axios.js
@@ -1,0 +1,3 @@
+import axios from 'axios';
+
+export default axios;
