+++ frontend/src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../lib/axios';
 import { useNavigate } from 'react-router-dom';
 
 const Home = () => {
