+++ frontend/src/pages/businesses/CreateBusiness.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../../lib/axios';
 import { toast } from 'react-toastify';
 import { Link, useNavigate } from 'react-router-dom';
 import { motion } from 'framer-motion';
