+++ frontend/src/pages/core/SystemSettings.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import axiosInstance from './lib/axios';
+import axiosInstance from '../../lib/axios';
 import { toast } from 'react-toastify';
 
 const SystemSettings = ({ isAdmin }) => {
